// TrashMob/client-app/src/components/Models/ContactTagData.ts
class ContactTagData {
    id: string = '00000000-0000-0000-0000-000000000000';
    name: string = '';
    color: string = '';
    createdByUserId: string | null = null;
    createdDate: string | null = null;
    lastUpdatedByUserId: string | null = null;
    lastUpdatedDate: string | null = null;
}

export default ContactTagData;